Update to deal with the debugging info in LLVM 3.0
Apparently the getCompileUnit() method in the debug info wrappers is
deprecated now (it asserts for the current debug info version).  This
change removes the calls to that and the associated fields in the debug
info structures.

Additionally, *some* other metadata references are allowed to be NULL.
I was not able to divine which it was, so now all are allowed to be
NULL (converted to Maybe in the Haskell bindings).