Make the bitcode parsing more scalable
For large bitcode files, the parser would just run out of memory.
Apparently the knot-tying code had enormous overhead in the map from C
pointers to Haskell values.  There are about 5 pointers worth of
overhead per key/value pair in a Map.  The HashMap in
unordered-containers is even worse.

This change switches to an IO hash table which is much faster and
lower overhead.  The tradeoff is that some unsafePerformIO black magic
is required to do the final knot tying.  It is actually safe and works fine
(unless bugs are introduced into the knot tying).

It is safe because all instructions are processed before any of the
tied-off values are demanded.  If this ceases to be the case, the code
will throw a KnotTyingException.

It might be worth switching the type and metadata translation over,
too.  Then the State monad could be ripped out and replaced with a
Reader.