Reorder type tests, fix handling of array metadata.
composite types are a subtype of derived types, so we need to check
for composite-ness first (otherwise we always get derived).

There seems to be no way to check a DIDescriptor to see if it is an
array.  We know which contexts arrays appear in, so just directly
translate those into arrays.