Match the LLVM-generated instruction labeling again.
For a while, this code gave names to void function calls where LLVM
did not.  This caused the labeling to drift out of sync and comparing
output from analysis phases to output from opt was difficult.

The recent changes to how Types are constructed allow us to do the
right thing here finally.